Commit 1 - Refactored to make Write and Read Benchmarks as separate classes, inheriting from a Benchmark superclass.
