+++ app/src/main/java/edu/touro/mcon152/bm/DiskMark.java
@@ -14,7 +14,7 @@ public class DiskMark {
     private double cumMin = 0;
     private double cumMax = 0;
     private double cumAvg = 0;
-    DiskMark(MarkType type) {
+    public DiskMark(MarkType type) {
         this.type = type;
     }
 
@@ -23,19 +23,19 @@ public class DiskMark {
         return "Mark(" + type + "): " + getMarkNum() + " bwMbSec: " + getBwMbSecAsString() + " avg: " + getAvgAsString();
     }
 
-    String getBwMbSecAsString() {
+    public String getBwMbSecAsString() {
         return df.format(getBwMbSec());
     }
 
-    String getMinAsString() {
+    public String getMinAsString() {
         return df.format(getCumMin());
     }
 
-    String getMaxAsString() {
+    public String getMaxAsString() {
         return df.format(getCumMax());
     }
 
-    String getAvgAsString() {
+    public String getAvgAsString() {
         return df.format(getCumAvg());
     }
