+++ app/src/main/java/edu/touro/mcon152/bm/DiskWorker.java
@@ -1,5 +1,8 @@
 package edu.touro.mcon152.bm;
 
+import edu.touro.mcon152.bm.benchmark.BenchmarkBase;
+import edu.touro.mcon152.bm.benchmark.ReadBenchmark;
+import edu.touro.mcon152.bm.benchmark.WriteBenchmark;
 import edu.touro.mcon152.bm.persist.DiskRun;
 import edu.touro.mcon152.bm.persist.EM;
 import jakarta.persistence.EntityManager;
@@ -36,7 +39,7 @@ public class DiskWorker {
     // Record any success or failure status returned from SwingWorker (might be us or super)
     Boolean lastStatus = true;  // so far unknown
     I_UI userInterface;
-    boolean status = true;
+    BenchmarkBase benchmark;
 
     /**
      * This gets the user interface for use in DIP.
@@ -76,129 +79,20 @@ public class DiskWorker {
                 + ", blk size (kb): " + App.blockSizeKb + ", blockSequence: " + App.blockSequence);
 
         /*
-          init local vars that keep track of benchmarks, and a large read/write buffer
-         */
-        int wUnitsComplete = 0, rUnitsComplete = 0, unitsComplete;
-        int wUnitsTotal = App.writeTest ? numOfBlocks * numOfMarks : 0;
-        int rUnitsTotal = App.readTest ? numOfBlocks * numOfMarks : 0;
-        int unitsTotal = wUnitsTotal + rUnitsTotal;
-        float percentComplete;
-
-        int blockSize = blockSizeKb * KILOBYTE;
-        byte[] blockArr = new byte[blockSize];
-        initializeBuffer(blockArr);
-
-        DiskMark wMark, rMark;  // declare vars that will point to objects used to pass progress to UI
-
-
-
-        int startFileNum = App.nextMarkNumber;
-
-        /*
           The DiskWriter allows a Write, Read, or both types of BMs to be started. They are done serially.
          */
-        if (App.writeTest) {
-            DiskRun run = new DiskRun(DiskRun.IOMode.WRITE, blockSequence);
-            setRunInfo(run);
-
-            // Tell logger and GUI to display what we know so far about the Run
-            msg("disk info: (" + run.getDiskInfo() + ")");
-            userInterface.setTitle(run);
-            // Create a test data file using the default file system and config-specified location
-            if (!App.multiFile) {
-                testFile = new File(dataDir.getAbsolutePath() + File.separator + "testdata.jdm");
+        if (writeTest){
+            benchmark = new WriteBenchmark(userInterface);
+            lastStatus = benchmark.runBenchmark();
+            if (!lastStatus){
+                return false;
             }
-
-            /*
-              Begin an outer loop for specified duration (number of 'marks') of benchmark,
-              that keeps writing data (in its own loop - for specified # of blocks). Each 'Mark' is timed
-              and is reported to the GUI for display as each Mark completes.
-             */
-            for (int m = startFileNum; m < startFileNum + App.numOfMarks && !hasBeenCancelled(); m++) {
-
-                if (App.multiFile) {
-                    testFile = new File(dataDir.getAbsolutePath()
-                            + File.separator + "testdata" + m + ".jdm");
-                }
-                wMark = new DiskMark(WRITE);    // starting to keep track of a new benchmark
-                wMark.setMarkNum(m);
-                long startTime = System.nanoTime();
-                long totalBytesWrittenInMark = 0;
-
-                String mode = "rw";
-                if (App.writeSyncEnable) {
-                    mode = "rwd";
-                }
-
-                try {
-                    try (RandomAccessFile rAccFile = new RandomAccessFile(testFile, mode)) {
-                        for (int b = 0; b < numOfBlocks; b++) {
-                            if (App.blockSequence == DiskRun.BlockSequence.RANDOM) {
-                                int rLoc = Util.randInt(0, numOfBlocks - 1);
-                                rAccFile.seek((long) rLoc * blockSize);
-                            } else {
-                                rAccFile.seek((long) b * blockSize);
-                            }
-                            rAccFile.write(blockArr, 0, blockSize);
-                            totalBytesWrittenInMark += blockSize;
-                            wUnitsComplete++;
-                            unitsComplete = rUnitsComplete + wUnitsComplete;
-                            percentComplete = (float) unitsComplete / (float) unitsTotal * 100f;
-
-                            /*
-                              Report to GUI what percentage level of Entire BM (#Marks * #Blocks) is done.
-                             */
-                            userInterface.setRunProgress((int) percentComplete);
-                        }
-                    }
-                } catch (IOException ex) {
-                    Logger.getLogger(App.class.getName()).log(Level.SEVERE, null, ex);
-                    lastStatus = false;
-                    return false;
-                }
-
-                /*
-                  Compute duration, throughput of this Mark's step of BM
-                 */
-                long endTime = System.nanoTime();
-                long elapsedTimeNs = endTime - startTime;
-                double sec = (double) elapsedTimeNs / (double) 1000000000;
-                double mbWritten = (double) totalBytesWrittenInMark / (double) MEGABYTE;
-                wMark.setBwMbSec(mbWritten / sec);
-                msg("m:" + m + " write IO is " + wMark.getBwMbSecAsString() + " MB/s     "
-                        + "(" + Util.displayString(mbWritten) + "MB written in "
-                        + Util.displayString(sec) + " sec)");
-                App.updateMetrics(wMark);
-
-                /*
-                  Let the GUI know the interim result described by the current Mark
-                 */
-                userInterface.setResults(wMark);
-
-                // Keep track of statistics to be displayed and persisted after all Marks are done.
-                run.setRunMax(wMark.getCumMax());
-                run.setRunMin(wMark.getCumMin());
-                run.setRunAvg(wMark.getCumAvg());
-                run.setEndTime(new Date());
-            } // END outer loop for specified duration (number of 'marks') for WRITE benchmark
-
-            /*
-              Persist info about the Write BM Run (e.g. into Derby Database) and add it to a GUI panel
-             */
-            EntityManager em = EM.getEntityManager();
-            em.getTransaction().begin();
-            em.persist(run);
-            em.getTransaction().commit();
-
-            userInterface.displayRun(run);
         }
-
         /*
           Most benchmarking systems will try to do some cleanup in between 2 benchmark operations to
           make it more 'fair'. For example a networking benchmark might close and re-open sockets,
           a memory benchmark might clear or invalidate the Op Systems TLB or other caches, etc.
          */
-
         // try renaming all files to clear catch
         if (App.readTest && App.writeTest && !hasBeenCancelled()) {
             String s =  """
@@ -210,112 +104,21 @@ public class DiskWorker {
             String s1 = "Clear Disk Cache Now";
             userInterface.showBlockingMessage(s, s1);
         }
-
         // Same as above, just for Read operations instead of Writes.
         if (App.readTest) {
-            DiskRun run = new DiskRun(DiskRun.IOMode.READ, App.blockSequence);
-            setRunInfo(run);
-
-            msg("disk info: (" + run.getDiskInfo() + ")");
-
-            userInterface.setTitle(run);
-
-            for (int m = startFileNum; m < startFileNum + App.numOfMarks && !hasBeenCancelled(); m++) {
-
-                if (App.multiFile) {
-                    testFile = new File(dataDir.getAbsolutePath()
-                            + File.separator + "testdata" + m + ".jdm");
-                }
-                rMark = new DiskMark(READ);  // starting to keep track of a new benchmark
-                rMark.setMarkNum(m);
-                long startTime = System.nanoTime();
-                long totalBytesReadInMark = 0;
-
-                try {
-                    try (RandomAccessFile rAccFile = new RandomAccessFile(testFile, "r")) {
-                        for (int b = 0; b < numOfBlocks; b++) {
-                            if (App.blockSequence == DiskRun.BlockSequence.RANDOM) {
-                                int rLoc = Util.randInt(0, numOfBlocks - 1);
-                                rAccFile.seek((long) rLoc * blockSize);
-                            } else {
-                                rAccFile.seek((long) b * blockSize);
-                            }
-                            rAccFile.readFully(blockArr, 0, blockSize);
-                            totalBytesReadInMark += blockSize;
-                            rUnitsComplete++;
-                            unitsComplete = rUnitsComplete + wUnitsComplete;
-                            percentComplete = (float) unitsComplete / (float) unitsTotal * 100f;
-                            userInterface.setRunProgress((int) percentComplete);
-                        }
-                    }
-                } catch (FileNotFoundException ex) {
-                    lastStatus = false;
-                    Logger.getLogger(App.class.getName()).log(Level.SEVERE, null, ex);
-                    String emsg = "May not have done Write Benchmarks, so no data available to read." +
-                            ex.getMessage();
-                    userInterface.showMessage(emsg, "Unable to read");
-                    msg(emsg);
-                    return false;
-                }
-                long endTime = System.nanoTime();
-                long elapsedTimeNs = endTime - startTime;
-                double sec = (double) elapsedTimeNs / (double) 1000000000;
-                double mbRead = (double) totalBytesReadInMark / (double) MEGABYTE;
-                rMark.setBwMbSec(mbRead / sec);
-                msg("m:" + m + " READ IO is " + rMark.getBwMbSec() + " MB/s    "
-                        + "(MBread " + mbRead + " in " + sec + " sec)");
-                App.updateMetrics(rMark);
-                userInterface.setResults(rMark);
-
-                run.setRunMax(rMark.getCumMax());
-                run.setRunMin(rMark.getCumMin());
-                run.setRunAvg(rMark.getCumAvg());
-                run.setEndTime(new Date());
+            benchmark = new ReadBenchmark(userInterface);
+            lastStatus = benchmark.runBenchmark();
+            if (!lastStatus){
+                return false;
             }
-
-            /*
-              Persist info about the Read BM Run (e.g. into Derby Database) and add it to a GUI panel
-             */
-            EntityManager em = EM.getEntityManager();
-            em.getTransaction().begin();
-            em.persist(run);
-            em.getTransaction().commit();
-
-            userInterface.displayRun(run);
-
         }
         App.nextMarkNumber += App.numOfMarks;
-        return true;
-    }
-
-    /**
-     * This sets up the buffer for testing purposes.
-     * @param blockArr
-     */
-    private static void initializeBuffer(byte[] blockArr) {
-        for (int b = 0; b < blockArr.length; b++) {
-            if (b % 2 == 0) {
-                blockArr[b] = (byte) 0xFF;
-            }
-        }
-    }
-
-    /**
-     * This is in charge of all the setting sof the current run of the software.
-     * @param run
-     */
-    private static void setRunInfo(DiskRun run) {
-        run.setNumMarks(App.numOfMarks);
-        run.setNumBlocks(App.numOfBlocks);
-        run.setBlockSize(App.blockSizeKb);
-        run.setTxSize(App.targetTxSizeKb());
-        run.setDiskInfo(Util.getDiskInfo(dataDir));
+        return lastStatus;
     }
     public void finish(){
         userInterface.finish();
     }
     public Boolean getLastStatus() {
-        lastStatus = status;
         return lastStatus;
     }
 }
